LayoutTests/imported/w3c/web-platform-tests/trusted-types/trusted-types-createHTMLDocument-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'trusted-types'.

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'require-trusted-types-for'.

shouldpass

FAIL Trusted Type assignment is blocked. (document) assert_throws_js: function "_ => { doc.createElement("script").textContent = "2+2"; }" did not throw
FAIL Trusted Type instances created in the main doc can be used. (document) promise_test: Unhandled rejection with value: object "ReferenceError: Can't find variable: trustedTypes"
FAIL Trusted Type assignment is blocked. (createHTMLDocument) assert_throws_js: function "_ => { doc.createElement("script").textContent = "2+2"; }" did not throw
FAIL Trusted Type instances created in the main doc can be used. (createHTMLDocument) promise_test: Unhandled rejection with value: object "ReferenceError: Can't find variable: trustedTypes"
FAIL Trusted Type assignment is blocked. (DOMParser) Can't find variable: trustedTypes
FAIL Trusted Type instances created in the main doc can be used. (DOMParser) Can't find variable: trustedTypes
FAIL Trusted Type assignment is blocked. (XHR) assert_throws_js: function "_ => { doc.createElement("script").textContent = "2+2"; }" did not throw
FAIL Trusted Type instances created in the main doc can be used. (XHR) promise_test: Unhandled rejection with value: object "ReferenceError: Can't find variable: trustedTypes"
FAIL Install default policy. promise_test: Unhandled rejection with value: object "ReferenceError: Can't find variable: trustedTypes"
FAIL Default policy applies. (document) assert_equals: expected "shouldpass [default]" but got "shouldpass"
FAIL Default policy applies. (createHTMLDocument) assert_equals: expected "shouldpass [default]" but got "shouldpass"
FAIL Default policy applies. (DOMParser) Can't find variable: trustedTypes
FAIL Default policy applies. (XHR) assert_equals: expected "shouldpass [default]" but got "shouldpass"
